fix: user, auth 에러 수정 ver 1
UserReponseDto와 ProfileResponseDto => auth의 UserResponseDto로 통합
SecurityConfig에서 api/는 허용 x
회원가입 시 유저 정보 초기화 없음 -> user default profile 설정, category 입력 feature에 업데이트
register / login dto는 분리 -> loginDto 사용
header에 토큰이 Authorization으로 통일
AuthenticationPrincipal UserDetails

+++ src/main/java/swe/second/team_matching_server/domain/meeting/model/dto/MeetingMemberElement.java
@@ -16,7 +16,10 @@ import lombok.EqualsAndHashCode;
 public class MeetingMemberElement extends UserElement {
     private MeetingMemberRole role;
 
-    public MeetingMemberElement(String id, String name, String profileUrl, int attendenceScore, Major major, String studentId, String phoneNumber, List<String> features, MeetingMemberRole role) {
+    public MeetingMemberElement(String id, String name, String profileUrl, byte attendenceScore,
+            Major major, String studentId, String phoneNumber, List<String> features,
+            MeetingMemberRole role) {
+
         super(id, name, profileUrl, attendenceScore, major, studentId, phoneNumber, features);
         this.role = role;
     }
