+++ src/main/java/swe/second/team_matching_server/domain/user/model/dto/UserElement.java
@@ -16,22 +16,17 @@ public class UserElement {
     private String id;
     private String name;
     private String profileUrl;
-    private int attendanceScore;
+    private byte attendanceScore;
     private Major major;
     private String studentId;
     private String phoneNumber;
     private List<String> features;
 
     public static UserElement of(User user) {
-        return UserElement.builder()
-                .id(user.getId())
-                .name(user.getUsername())
+        return UserElement.builder().id(user.getId()).name(user.getUsername())
                 .profileUrl(user.getProfileImage() != null ? user.getProfileImage().getUrl() : null)
-                .attendanceScore(user.getAttendanceScore())
-                .major(user.getMajor())
-                .studentId(user.getStudentId())
-                .phoneNumber(user.getPhoneNumber())
-                .features(user.getFeatures())
-                .build();
+                .attendanceScore(user.getAttendanceScore()).major(user.getMajor())
+                .studentId(user.getStudentId()).phoneNumber(user.getPhoneNumber())
+                .features(user.getFeatures()).build();
     }
 }
